Move Migrations from 1.10.1 to 1.10.5 for Workflows
