+++ openmetadata-service/src/main/java/org/openmetadata/service/migration/utils/v1105/MigrationUtil.java
@@ -1,4 +1,4 @@
-package org.openmetadata.service.migration.utils.v1101;
+package org.openmetadata.service.migration.utils.v1105;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
